backend/requirement.txt
openai
flask
python-dotenv
requests
gunicorn
